src/requirements.txt
flask>=3.1.2,<4.0.0
numpy>=2.3.4,<3.0.0
gunicorn>=23.0.0,<24.0.0
pandas>=2.3.3,<3.0.0
cantools>=40.7.1,<41.0.0
rerun-sdk==0.24.1